+++ src/read_shared_memory_example.cpp
@@ -27,7 +27,7 @@ struct positrack_shared_memory
   double x[POSITRACKSHARENUMFRAMES]; // position x
   double y[POSITRACKSHARENUMFRAMES]; // position y
   double hd[POSITRACKSHARENUMFRAMES]; // head direction
-  int trialNo[POSITRACKSHARENUMFRAMES]; // current trial number
+  unsigned long int trialNo[POSITRACKSHARENUMFRAMES]; // current trial number
   pthread_mutexattr_t attrmutex;
   int is_mutex_allocated;
   pthread_mutex_t pmutex;
